requirements.txt
fastapi==0.112.1
python-socketio==5.11.3
uvicorn==0.30.6
